Still a bit flickery when video running and position_changes are forced -> open timeouts emit outdated positions

+++ src/media_backend/type_specific_player/mocap.py
@@ -41,7 +41,7 @@ class MocapLoader(AbstractMediaLoader):
                 
                 frames[frame_index, :, :] = frame
 
-            self.media = frames.astype(np.float16)
+            self.media = frames.astype(np.float32)
             # self.media = frames
             
         except Exception as e:
